#11 - A little more useful information and methods added to requests, responses, logging and toasts...
Now deleting data files from device if upload succeeds (it doesn't care what the destination does with the data, only that the upload succeeded. SRP ftw!)
Naiive example of a Web API controller method which can deserialise the data is provided.
Yes! It does need a shave and a haircut!